Controller decomposed - code refactored
